build - fix overall sysconfdir handling so it can work out of the box
sysconfdir is "etc" by default in meson, this means paths are like
"etc/PackageKit/xxxx" ... which is not a path. missing a / to begin
with. not to mention packagekit installs its own etc files in prefix/etc
(eg /usr/local/etc) by default. so differentiate files installed by
pkgkit (pk_sysconfdir)) and other system files (sysconfdir) and use
approplriately in the build. if the default etc is used then fix it to be
/etc for system and prefix/etc for pk installed files. at least things can
work out of the box then when you ninja install and dont provide a
custom sysconfidr on the cmdline.